requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
defusedxml==0.5.0
Django==2.0.5
django-filter==1.1.0
django-jsonfield==1.0.1
django-notify-x==0.1.9
djangorestframework==3.8.2
gunicorn==19.8.1
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
lxml==4.2.1
Markdown==2.6.11
oauthlib==2.1.0
parsel==1.4.0
Pillow==5.1.0
pipenv==2018.5.18
psycopg2==2.7.4
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
PyJWT==1.6.4
pyOpenSSL==18.0.0
python3-openid==3.1.0
pytz==2018.4
queuelib==1.5.0
requests==2.18.4
requests-oauthlib==1.0.0
selenium==3.12.0
service-identity==17.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
w3lib==1.19.0
whitenoise==3.3.1
zope.interface==4.5.0
